+++ packages/eas-cli/src/commands/simulator/index.ts
@@ -38,6 +38,7 @@ import {
   formatRemoteSessionInstructions,
   formatSimulatorUnavailableMessage,
   getRemoteSessionEnvironmentVariables,
+  remoteConfigWithPreviewPageUrl,
 } from '../../simulator/utils';
 import { enableJsonOutput, printJsonOnlyOutput } from '../../utils/json';
 import { sleepAsync } from '../../utils/promise';
@@ -385,7 +386,7 @@ export default class Simulator extends EasCommand {
         name,
         type: flags.type,
         deviceRunSessionUrl,
-        remoteConfig,
+        remoteConfig: remoteConfigWithPreviewPageUrl(remoteConfig),
       });
       return;
     }